+++ src/main/java/com/triptune/global/config/CorsConfig.java
@@ -1,0 +1,38 @@
+package com.triptune.global.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.List;
+
+@Configuration
+@ConfigurationProperties(prefix = "cors")
+public class CorsConfig {
+    private List<String> allowedOrigins;
+
+    public List<String> getAllowedOrigins(){
+        return allowedOrigins;
+    }
+
+    public void setAllowedOrigins(List<String> allowedOrigins){
+        this.allowedOrigins = allowedOrigins;
+    }
+
+
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource(){
+        CorsConfiguration configuration = new CorsConfiguration();
+        configuration.setAllowedOrigins(allowedOrigins);
+        configuration.setAllowedMethods(List.of("GET", "POST", "PATCH", "DELETE"));
+        configuration.setAllowedHeaders(List.of("Content-Type", "Authorization"));
+        configuration.setAllowCredentials(true);
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration); // 모든 경로에 적용
+        return source;
+    }
+}
